Update parsing_fc_color_two.c: change variable type from int to long long for RGB value extraction

+++ src/parsing/parsing_fc_color_two.c
@@ -6,7 +6,7 @@
 /*   By: jariskan <jariskan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/11 16:45:52 by smarin-a          #+#    #+#             */
-/*   Updated: 2025/03/21 02:59:33 by jariskan         ###   ########.fr       */
+/*   Updated: 2025/03/21 17:05:57 by jariskan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 static int	ft_extract_rgb_value(char **color_str)
 {
-	int	value;
-	int	digits;
+	long long	value;
+	int			digits;
 
 	value = 0;
 	digits = 0;
